Calculate Average Wear Time
-add more member variables.
-add the Average Wear Time number TextView.
-update explanation of how to calculate the Last five Wear Times.
-add code to tell android what to do when the user presses the
reset button.
-When the user presses the reset button it will shift the numbers
in the Last Five Wear Times to the right by one slot. It will also take
note of the number noted in the slot and save it as both and integer
and a string.
-complete code that will calculate the Average Wear Time based on
if the slot has a zero in it or a number greater that zero.  If a particular
slot has a zero in it then it will most likely mean that the user has been
using the app less than 10 weeks (on average).  Typically all Last Five
Wear Times slots won't have a number in it until the user has been
using the app for 10 weeks.  This is based on the average wear time
of 2 weeks per pair of contact lens.
-add String for error message toast