Complete shared preferences
-complete set up of shared preferences in the
resetCurrentDaysWorn method.
-complete set up of shared preferences in the
calculateAverageWearTime method.  This will add the
values of the last five wear time slots with values other
than zero and the average wear time number into the
shared prefs.
-complete set up shared preferences in the
isTheUserNew method.  If the user is new it will
create all the shared prefs and give them a value of zero.