Contact lens prescription shared preferences
-add shared prefs for the contacts prescription
and glasses prescription.
-add boolean and pref to determine if the user
is new.
-complete saveContactsPrescription method.  Perform
a check to see if the user is new.  If they are new, add
generic values to the shared prefs. Also set the
generic info on the respective edit text views.
If the user is not new then get the saved info
from the shared prefs and set them on their
respective edit text views.