Add Last Worn Date to Main Activity
-increment targetSdk from 31 to 32.
-add two text views to activity_main.xml.
I added a text view for the title Last Worn
and another text view for the actual last worn
date.
-add code in MainActivity to find the Last Worn
text views and the preferences.
-add code to IsTheUserNew method to add
the last worn date into the shared prefs, and
also to get the last worn date out of the shared
prefs in case the user is not new.
-add strings and dimens for the last worn
actual dates.