Fix checkForColorChange method
-add code in the checkForColorChange method
to get the mCurrentLensCountString
and parse it into an Integer before checking the
value of the Integer so that we can set the correct
color in the Main Wear Status Color textview.
This corrects the problem of the Main Wear
Color Status displaying the wrong color after
the user resets the Lens Counter and then
closes the app. When they re-enter the app
it would display the previous color status
and not the correct color status.
-update app dependencies.
-change the app version number from 1.3 to
version 1.4 in strings.xml.